Small Update
I have reached a point where I don't know how to implement what is left to do. This includes the ability to add an item to cart, create a product page for each item clicked from the gallery. Also, how to use the square API to pull item data and to do the checkout implementation. Until I learn to do those things or get help with the development, this project is at a stalemate.